GUI_indian railway/requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
pygame==2.0.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
xlrd==1.2.0
